moved pull to refresh to BasicDataStateUi
