+++ feature/course-view/src/debug/kotlin/de/tum/informatics/www1/artemis/native_app/feature/courseview/MessagingScreenshots.kt
@@ -146,8 +146,7 @@ fun `Metis - Conversation Overview`() {
                     onRequestCreatePersonalConversation = {},
                     onRequestAddChannel = {},
                     onRequestBrowseChannel = {},
-                    canCreateChannel = false,
-                    onRefresh = {},
+                    canCreateChannel = false
                 )
             },
             onNavigateBack = { },
