+++ feature/course-view/src/main/kotlin/de/tum/informatics/www1/artemis/native_app/feature/courseview/ui/WeeklyItemsUi.kt
@@ -10,6 +10,7 @@ import androidx.compose.foundation.layout.asPaddingValues
 import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.foundation.layout.padding
 import androidx.compose.foundation.layout.systemBars
+import androidx.compose.foundation.lazy.LazyColumn
 import androidx.compose.foundation.lazy.items
 import androidx.compose.material.icons.Icons
 import androidx.compose.material.icons.filled.ExpandLess
@@ -28,7 +29,6 @@ import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.unit.dp
-import de.tum.informatics.www1.artemis.native_app.core.ui.compose.RefreshableLazyColumn
 import de.tum.informatics.www1.artemis.native_app.feature.courseview.GroupedByWeek
 import de.tum.informatics.www1.artemis.native_app.feature.courseview.R
 import kotlinx.datetime.Clock
@@ -51,7 +51,6 @@ internal fun <T> WeeklyItemsLazyColumn(
     modifier: Modifier,
     weeklyItemGroups: List<GroupedByWeek<T>>,
     verticalArrangement: Arrangement.Vertical = Arrangement.spacedBy(8.dp),
-    onRefresh: () -> Unit,
     getItemId: T.() -> Long,
     itemContent: @Composable (T) -> Unit
 ) {
@@ -76,13 +75,12 @@ internal fun <T> WeeklyItemsLazyColumn(
         }
     }
 
-    RefreshableLazyColumn(
+    LazyColumn(
         modifier = modifier,
         verticalArrangement = verticalArrangement,
         contentPadding = PaddingValues(
             bottom = WindowInsets.systemBars.asPaddingValues().calculateBottomPadding()
-        ),
-        onRefresh = onRefresh
+        )
     ) {
         weeklyItemGroups.forEachIndexed { index, weeklyItems ->
             item {
