+++ feature/dashboard/src/main/kotlin/de/tum/informatics/www1/artemis/native_app/feature/dashboard/ui/CourseList.kt
@@ -46,13 +46,11 @@ import java.text.DecimalFormat
 fun CourseList(
     modifier: Modifier,
     courses: List<CourseWithScore>,
-    onRefresh: () -> Unit,
     onClickOnCourse: (Course) -> Unit
 ) {
     CourseItemGrid(
         modifier = modifier,
         courses = courses,
-        onRefresh = onRefresh,
     ) { dashboardCourse, courseItemModifier, isCompact ->
         CourseItem(
             modifier = courseItemModifier.testTag(testTagForCourse(dashboardCourse.course.id!!)),
